+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 10. [StartScreen] Resize fix x86.    (Credits to [Slaynash](https://github.com/Slaynash) :3)
 11. [StartScreen] Fixed crash on mono.    (Credits to [Slaynash](https://github.com/Slaynash) :3)
 12. Rewrote MelonLaunchOptions Loading.
+13. Temporarily Reverted Il2Cpp Coroutine Interface changes.
 
 ---